Minor file restructuring
