+++ internal/help_class.py
